+++ NEWS.md
@@ -4,7 +4,8 @@
 
 * Improved plotting: now plotting the GAM-smoothed signal with raw data and improved clusters aesthetics.
 * Allowing 3 different models to be fitted: full GAMM, GAMM with summary statistics (recommended), or group-level GAM.
-* Adding further support for presence or absence of predictor (e.g., group, condition). When predictor_id=NA, neurogam now tests signal against 0 through time.
+* Adding further support for presence or absence of predictor (e.g., group, condition). When `predictor_id = NA`, `neurogam` now tests signal against 0 through time.
+* Implementing `print()` and `summary()` methods for `cluster_results` objects.
 
 ## Other changes